+++ ui/src/components/comments/list/item/footer/vote/index.ts
