+++ ui/src/components/comments/list/item/header/index.ts
@@ -1,1 +1,2 @@
 export { ListItemTimestamp } from "./timestamp";
+export { CommentWebsiteInfo } from "./website";
