+++ ui/src/components/comments/list/item/index.ts
@@ -1,3 +1,3 @@
-export { ListItemTimestamp } from "./header";
+export { ListItemTimestamp, CommentWebsiteInfo } from "./header";
 export { ListItemContent } from "./content";
 export { ListItemFooter } from "./footer";
